requirements.txt
flask==2.3.3
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
Werkzeug==2.3.7
pdf2image>=1.16.3
